requirements.txt
python-dotenv
slack-bolt
tiktoken
requests
pillow
openai
beautifulsoup4
tqdm